% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-get_budget.R
\name{get_budgets}
\alias{get_budgets}
\title{get_budgets}
\usage{
get_budgets(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI budgets, this can be
`iati_identifier`, `title`, `description`, `recipient_country`,
`recipient_region`, `recipient_org`, `sector`, `document_link`,
`participating_org`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on budgets from IATI-compliant datasets.
}
\description{
Function to get budgets endpoint from IATI-compliant datasets.
}
\examples{
get_budgets(url = "https://devtracker.dfid.gov.uk",
            query = "iati_identifier",
            value = "075004")

}
