% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-get_cities.R
\name{get_cities}
\alias{get_cities}
\title{get_cities}
\usage{
get_cities(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI activites, this can be
specified as `city_id`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on cities from IATI-compliant datasets.
}
\description{
Function to get cities endpoint from IATI-compliant datasets.
}
\examples{
get_cities(url = "https://devtracker.dfid.gov.uk",
           query = "city_id",
           value = "3443013")

}
