% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-get_countries.R
\name{get_countries}
\alias{get_countries}
\title{get_countries}
\usage{
get_countries(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI countries, this can be
specified as `country_id`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on countries from IATI-compliant datasets.
}
\description{
Function to get activities endpoint from IATI-compliant datasets.
}
\examples{
get_countries(url = "https://devtracker.dfid.gov.uk",
              query = "country_id",
              value = "AC")

}
