% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{get_datasets}
\usage{
get_datasets(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI datasets, this can be
specified as `name`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on datasets from IATI-compliant datasets.
}
\description{
Function to get datasets endpoint from IATI-compliant datasets.
}
\examples{
get_datasets(url = "https://devtracker.dfid.gov.uk",
             query = "name",
             value = "6236")

}
