% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-get_locations.R
\name{get_locations}
\alias{get_locations}
\title{get_locations}
\usage{
get_locations(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI activites, this can be
specified as `activity_id`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on locations from IATI-compliant datasets.
}
\description{
Function to get locations endpoint from IATI-compliant datasets.
}
\examples{
get_locations(url = "https://devtracker.dfid.gov.uk",
              query = "activity_id",
              value = "883116")

}
