% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oipa_api.R
\name{get_oipa_activities}
\alias{get_oipa_activities}
\title{get_oipa_activities}
\usage{
get_oipa_activities(base, iati_identifier, query = NULL)
}
\arguments{
\item{base}{Base URL}

\item{iati_identifier}{IATI unique identifier}

\item{query}{Query to pass to API}
}
\value{
Parsed content
}
\description{
Function to get activities from IATI-compliant sites.
}
\examples{
#

}
