% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_organisations.R
\name{get_organisations}
\alias{get_organisations}
\title{get_organisations}
\usage{
get_organisations(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI organisations, this can is
specified as `organisation_id`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on organisations from IATI-compliant datasets.
}
\description{
Function to get organisations endpoint from IATI-compliant datasets.
}
\examples{
get_organisations(url = "https://devtracker.dfid.gov.uk",
                  query = "organisation_id",
                  value = "US-EIN-262681792")

}
