% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_publishers.R
\name{get_publishers}
\alias{get_publishers}
\title{get_publishers}
\usage{
get_publishers(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI publishers, this can be
`id`, `publisher_iati_id`, `display_name`, `name`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on publishers from IATI-compliant datasets.
}
\description{
Function to get publishers endpoint from IATI-compliant datasets.
}
\examples{
get_publishers(url = "https://devtracker.dfid.gov.uk",
               query = "publisher_iati_id",
               value = "075004")

}
