% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-get_regions.R
\name{get_regions}
\alias{get_regions}
\title{get_regions}
\usage{
get_regions(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI activites, this can be
specified as `region_id`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on regions from IATI-compliant datasets.
}
\description{
Function to get regions endpoint from IATI-compliant datasets.
}
\examples{
get_regions(url = "https://devtracker.dfid.gov.uk",
            query = "region_id",
            value = "189")

}
