% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-get_results.R
\name{get_results}
\alias{get_results}
\title{get_results}
\usage{
get_results(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI activites, this can be
`iati_identifier`, `title`, `description`, `recipient_country`,
`recipient_region`, `recipient_org`, `sector`, `document_link`,
`participating_org`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on results from IATI-compliant datasets.
}
\description{
Function to get results endpoint from IATI-compliant datasets.
}
\examples{
get_results(url = "https://devtracker.dfid.gov.uk",
            query = "iati_identifier",
            value = "075004")

}
