% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-get_sectors.R
\name{get_sectors}
\alias{get_sectors}
\title{get_sectors}
\usage{
get_sectors(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI activites, this can be
specified as `sector_id`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on sectors from IATI-compliant datasets.
}
\description{
Function to get sectors endpoint from IATI-compliant datasets.
}
\examples{
get_sectors(url = "https://devtracker.dfid.gov.uk",
            query = "sector_id",
            value = "11110")

}
