% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-get_transactions.R
\name{get_transactions}
\alias{get_transactions}
\title{get_transactions}
\usage{
get_transactions(url, query, value)
}
\arguments{
\item{url}{URL for websites containing IATI-compliant datasets.}

\item{query}{Search/query parameter. For IATI activites, this can be
`description`, `provider_organisation_name`, `receiver_organisation_name`.}

\item{value}{Value to pass to specified `query`.}
}
\value{
Parsed content on transactions from IATI-compliant datasets.
}
\description{
Function to get transactions endpoint from IATI-compliant datasets.
}
\examples{
get_transactions(url = "https://devtracker.dfid.gov.uk",
                 query = "transaction_id",
                 value = "270304")

}
