% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oipa.R
\docType{package}
\name{oipa}
\alias{oipa}
\alias{oipa-package}
\title{oipa}
\description{
An R Interface to the Openaid International Aid Transparency Initiative
Parser and API
}
\details{
The International Aid Transparency Initiatives (IATI) is a global aid
transparency standard and it makes information about aid spending easier to
access, use and understand using a unified open standard. Openaid IATI Parser
and Application Programming Interface (OIPA) enables IATI standard compliant
dadtasets to easily parse and interface that data using API that has been
developed on the Django framework. This package provides an R interface to
the OIPA.
}
